refactor(filter/map/reduce): refactor method mapKeys_ on mutable objects
It's been hard to refactor this. During the process, a bug was introduced. There is a comment
pointing out some details on that bug:  //applying the function before doing any update in the
object is essential. It's related to side effects and mutability and in those scenarios the order
matters.

re #11